공지 팝업 위치 변경
